[GFLAGS] Move code around to handle multiple options
