+++ modules/rosapps/applications/cmdutils/gflags/CMakeLists.txt
@@ -1,7 +1,14 @@
 
-add_executable(gflags gflags.c gflags.rc)
+
+list(APPEND SOURCE
+    gflags.c
+    pageheap.c
+    gflags.h)
+
+add_executable(gflags ${SOURCE} gflags.rc)
 set_module_type(gflags win32cui UNICODE)
 add_importlibs(gflags advapi32 user32 msvcrt kernel32)
+add_pch(gflags gflags.h SOURCE)
 
 if(MSVC)
     add_importlibs(gflags ntdll)
